Auth@Claim

+++ API/Services/CharacterService/CharacterService.cs
@@ -3,6 +3,7 @@ using API.Dtos.Character;
 using API.Models;
 using AutoMapper;
 using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
 
 namespace API.Services.CharacterService
 {
@@ -10,22 +11,32 @@ namespace API.Services.CharacterService
     {
         private readonly IMapper _mapper;
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             _mapper = mapper;
             _context = context;
+            _httpContextAccessor = httpContextAccessor;
         }
+
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext!.User
+            .FindFirstValue(ClaimTypes.NameIdentifier)!);
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacterAsync(AddCharacterDto newCharacter)
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             var character = _mapper.Map<Character>(newCharacter);
+            character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
             
-            await _context.Characters.AddAsync(_mapper.Map<Character>(newCharacter));
+            await _context.Characters.AddAsync(character);
             await _context.SaveChangesAsync();
 
-            serviceResponse.Data = await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+
+            serviceResponse.Data = await _context.Characters
+                .Where(c => c.User!.Id == GetUserId())
+                .Select(c => _mapper.Map<GetCharacterDto>(c))
+                .ToListAsync();
 
             return serviceResponse;
         }
@@ -36,14 +47,17 @@ namespace API.Services.CharacterService
             try
             {
 
-                var character = await _context.Characters.SingleOrDefaultAsync(c => c.Id == id);
+                var character = await _context.Characters.SingleOrDefaultAsync(c => c.Id == id );
 
-                if (character is null) throw new Exception($"Character with Id '{id}'Not Found");
+                if (character is null || character.User!.Id == GetUserId()) throw new Exception($"Character with Id '{id}'Not Found");
 
                 _context.Characters.Remove(character);
                 await _context.SaveChangesAsync();
 
-                serviceResponse.Data = await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+                serviceResponse.Data = await _context.Characters
+                    .Where(c => c.User!.Id == GetUserId())
+                    .Select(c => _mapper.Map<GetCharacterDto>(c))
+                    .ToListAsync();
 
 
             }
@@ -59,7 +73,7 @@ namespace API.Services.CharacterService
         public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharactersAsync()
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            var dbCharacters = await _context.Characters.ToListAsync();
+            var dbCharacters = await _context.Characters.Where(c => c.User!.Id == GetUserId()).ToListAsync();
             serviceResponse.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
             return serviceResponse;
         }
@@ -67,7 +81,14 @@ namespace API.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterByIdAsync(int id)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbCharacter = await _context.Characters.SingleOrDefaultAsync(c => c.Id == id);
+            var dbCharacter = await _context.Characters.SingleOrDefaultAsync(c => c.Id == id && c.User!.Id == GetUserId());
+
+            if (dbCharacter == null)
+            {
+                serviceResponse.Success = false;
+                serviceResponse.Massage = "Noy Found";
+                return serviceResponse;
+            }
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
             
             if (serviceResponse == null) throw new Exception("Character not found");
@@ -81,9 +102,11 @@ namespace API.Services.CharacterService
             try 
             {
                 
-                var character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+                var character = await _context.Characters
+                    .Include(u => u.User)
+                    .FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
 
-                if (character == null) throw new Exception($"Character with Id '{updatedCharacter.Id}'Not Found");
+                if (character == null || character.User!.Id == GetUserId()) throw new Exception($"Character with Id '{updatedCharacter.Id} 'Not Found");
 
                 character.Name = updatedCharacter.Name;
                 character.HitPoints = updatedCharacter.HitPoints;
